+++ app/javascript/components/table/Row.vue
@@ -2,7 +2,10 @@
 <tr>
     <td>{{ item.name }}</td>
     <td>{{ item.designation }}</td>
-    <td><a :href="wdpaUrl(item.wdpa_id)" title="View protected area on Protected Planet" target="_blank">{{ item.wdpa_id }}</a></td>
+    <td>
+      <template v-if="item.restricted">{{ item.wdpa_id }}</template>
+      <a v-else :href="wdpaUrl(item.wdpa_id)" title="View protected area on Protected Planet" target="_blank">{{ item.wdpa_id }}</a>
+    </td>
     <td>{{ item.id }}</td>
     <td>{{ checkForMultiples('iso3') }}</td>
     <td>{{ item.methodology }}</td>
@@ -32,7 +35,7 @@
 
     methods: {
       wdpaUrl (wdpaId) {
-        return 'https://protectedplanet.net/' + wdpaId
+        return `https://protectedplanet.net/${wdpaId}`
       },
 
       assessmentUrl (url) {