+++ app/models/pame_evaluation.rb
@@ -1,10 +1,11 @@
 require 'csv'
 
 class PameEvaluation < ApplicationRecord
-  belongs_to :protected_area
+  belongs_to :protected_area, optional: true
   belongs_to :pame_source
+  has_and_belongs_to_many :countries
 
-  validates :methodology, :year, :protected_area, :metadata_id, :url, presence: true
+  validates :methodology, :year,:metadata_id, :url, presence: true
 
   TABLE_ATTRIBUTES = [
     {
@@ -17,7 +18,8 @@ class PameEvaluation < ApplicationRecord
     },
     {
       title: 'WDPA ID',
-      field: 'wdpa_id'
+      field: 'wdpa_id',
+      tooltip: 'Unrestricted Protected Areas can be viewed on Protected Planet'
     },
     {
       title: 'Assessment ID',
@@ -61,15 +63,15 @@ class PameEvaluation < ApplicationRecord
     {
       current_page: page,
       per_page: 100,
-      total_entries: items.total_entries,
-      total_pages: items.total_pages,
+      total_entries: (items.count > 0 ? items[0][:total_entries] : 0),
+      total_pages:   (items.count > 0 ? items[0][:total_pages] : 0),
       items: items
     }
   end
 
   def self.generate_query(page, filter_params)
     # if params are empty then return the paginated results without filtering
-    return PameEvaluation.order('id ASC').paginate(page: page || 1, per_page: 100) if filter_params.empty?
+    return PameEvaluation.where(visible: true).order('id ASC').paginate(page: page || 1, per_page: 100) if filter_params.empty?
 
     filters = filter_params.select { |hash| hash["options"].present? }
 
@@ -79,19 +81,22 @@ class PameEvaluation < ApplicationRecord
 
   def self.parse_filters(filters)
     site_ids = []
-    where_params = {sites: "", methodology: "", year: ""}
+    country_ids = []
+    where_params = {sites: "", methodology: "", year: "", iso3: ""}
     filters.each do |filter|
       options = filter["options"]
       case filter['name']
       when 'iso3'
         countries = options
         site_ids << countries.map{ |iso3| Country.find_by(iso_3: iso3).protected_areas.pluck(:id) }
-        where_params[:sites] = options.empty? ? nil : "protected_area_id IN (#{site_ids.join(',')})"
+        where_params[:sites] = site_ids.flatten.empty? ? nil : "protected_area_id IN (#{site_ids.join(',')})"
+        country_ids << countries.map{ |iso3| "#{ Country.find_by(iso_3: iso3).id }" }
+        where_params[:iso3] = country_ids.flatten.empty? ? nil : "countries.id IN (#{country_ids.join(',')})"
       when 'methodology'
         options = options.map{ |e| "'#{e}'" }
-        where_params[:methodology] = options.empty? ? nil : "#{filter["name"]} IN (#{options.join(',')})"
+        where_params[:methodology] = options.empty? ? nil : "methodology IN (#{options.join(',')})"
       when 'year'
-        where_params[:year] = options.empty? ? nil : "#{filter["name"]} IN (#{options.join(',')})"
+        where_params[:year] = options.empty? ? nil : "year IN (#{options.join(',')})"
       end
     end
     where_params
@@ -99,11 +104,12 @@ class PameEvaluation < ApplicationRecord
 
   def self.run_query(page, where_params)
     if where_params[:sites].present?
+      query = PameEvaluation.connection.unprepared_statement {
+        "((#{pame_evaluations_from_pa_query(where_params)}) UNION (#{pame_evaluations_from_countries_query(where_params)})) AS pame_evaluations"
+      }
+
       PameEvaluation
-      .joins(:protected_area)
-      .where(where_params[:sites])
-      .where(where_params[:methodology])
-      .where(where_params[:year])
+      .from(query)
       .paginate(page: page || 1, per_page: 100).order('id ASC')
     else
       PameEvaluation
@@ -113,29 +119,54 @@ class PameEvaluation < ApplicationRecord
     end
   end
 
+  def self.pame_evaluations_from_pa_query(where_params)
+    PameEvaluation
+    .joins(:protected_area)
+    .where(where_params[:sites])
+    .where(where_params[:methodology])
+    .where(where_params[:year])
+    .to_sql
+  end
+
+  def self.pame_evaluations_from_countries_query(where_params)
+    PameEvaluation
+    .joins(:countries)
+    .where(where_params[:iso3])
+    .where(where_params[:methodology])
+    .where(where_params[:year])
+    .to_sql
+  end
+
   def self.serialise(evaluations)
-    evaluations.to_a.map! do |evaluation|
+    evaluations.select{|pe| pe.visible}.to_a.map! { |evaluation|
+      
+      wdpa_id = evaluation.protected_area&.wdpa_id || evaluation.wdpa_id
+      name  = evaluation.protected_area&.name || evaluation.name
+      designation = evaluation.protected_area&.designation&.name || "N/A"
+      countries = evaluation.protected_area&.countries || evaluation.countries
+      iso3 = countries.pluck(:iso_3).sort
       {
         current_page: evaluations.current_page,
         per_page: evaluations.per_page,
         total_entries: evaluations.total_entries,
         total_pages: evaluations.total_pages,
         id: evaluation.id,
-        wdpa_id: evaluation.protected_area.wdpa_id,
-        iso3: evaluation.protected_area.countries.pluck(:iso_3).sort,
+        wdpa_id: wdpa_id,
+        restricted: evaluation.restricted,
+        iso3: iso3,
         methodology: evaluation.methodology,
         year: evaluation.year.to_s,
         url: evaluation.url,
         metadata_id: evaluation.metadata_id,
-        source_id: evaluation.pame_source.id,
-        name: evaluation.protected_area.name,
-        designation: evaluation.protected_area.designation.name,
-        data_title: evaluation.pame_source.data_title,
-        resp_party: evaluation.pame_source.resp_party,
-        language: evaluation.pame_source.language,
-        source_year: evaluation.pame_source.year
+        source_id: evaluation.pame_source&.id,
+        name: name,
+        designation: designation,
+        data_title: evaluation.pame_source&.data_title,
+        resp_party: evaluation.pame_source&.resp_party,
+        language: evaluation.pame_source&.language,
+        source_year: evaluation.pame_source&.year
       }
-    end
+    }
   end
 
   def self.sources_to_json
@@ -178,8 +209,9 @@ class PameEvaluation < ApplicationRecord
     ].to_json
   end
 
-  def self.generate_csv(where_statement)
+  def self.generate_csv(where_statement, restricted_where_statement)
     where_statement = where_statement.empty? ? '' : "WHERE #{where_statement}"
+    restricted_where_statement = restricted_where_statement.empty? ? '' : "WHERE #{restricted_where_statement}"
     query = <<-SQL
       SELECT e.id AS id,
              e.metadata_id AS metadata_id,
@@ -196,12 +228,35 @@ class PameEvaluation < ApplicationRecord
              pame_sources.language AS language
              FROM pame_evaluations e
              INNER JOIN protected_areas ON e.protected_area_id = protected_areas.id
-             INNER JOIN pame_sources ON e.pame_source_id = pame_sources.id
+             LEFT JOIN pame_sources ON e.pame_source_id = pame_sources.id
              INNER JOIN countries_protected_areas ON protected_areas.id = countries_protected_areas.protected_area_id
              INNER JOIN countries ON countries_protected_areas.country_id = countries.id
              INNER JOIN designations ON protected_areas.designation_id = designations.id
              #{where_statement}
              GROUP BY e.id, protected_areas.wdpa_id, protected_areas.name, designation, pame_sources.data_title,
+                      pame_sources.resp_party, pame_sources.year, pame_sources.language
+
+             UNION
+
+      SELECT e.id AS id,
+             e.metadata_id AS metadata_id,
+             e.url AS url,
+             e.year AS evaluation_year,
+             e.methodology AS methodology,
+             0 AS wdpa_id,
+             ARRAY_TO_STRING(ARRAY_AGG(countries.iso_3),';') AS countries,
+             e.name AS site_name,
+             NULL AS designation,
+             pame_sources.data_title AS data_title,
+             pame_sources.resp_party AS resp_party,
+             pame_sources.year AS source_year,
+             pame_sources.language AS language
+             FROM pame_evaluations e
+             INNER JOIN pame_sources ON e.pame_source_id = pame_sources.id
+             INNER JOIN countries_pame_evaluations ON e.id = countries_pame_evaluations.pame_evaluation_id
+             INNER JOIN countries ON countries_pame_evaluations.country_id = countries.id
+             #{restricted_where_statement}
+             GROUP BY e.id, wdpa_id, e.name, designation, pame_sources.data_title,
                       pame_sources.resp_party, pame_sources.year, pame_sources.language;
     SQL
     evaluations = ActiveRecord::Base.connection.execute(query)
@@ -211,28 +266,29 @@ class PameEvaluation < ApplicationRecord
       evaluation_columns = PameEvaluation.new.attributes.keys
       evaluation_columns << "evaluation_id"
 
-      excluded_attributes = ["protected_area_id", "pame_source_id", "created_at", "updated_at", "id", "site_id", "source_id"]
+      excluded_attributes = ["visible", "restricted", "protected_area_id", "pame_source_id", "created_at", "updated_at", "id", "site_id", "source_id"]
 
       evaluation_columns.delete_if { |k, v| excluded_attributes.include? k }
 
-      additional_columns = ["wdpa_id", "iso3", "name", "designation", "source_data_title", "source_resp_party", "source_year", "source_language"]
+      additional_columns = ["iso3", "designation", "source_data_title", "source_resp_party", "source_year", "source_language"]
       evaluation_columns << additional_columns.map{ |e| "#{e}" }
 
       csv_line << evaluation_columns.flatten
 
       evaluations.each do |evaluation|
         evaluation_attributes = PameEvaluation.new.attributes
+
         evaluation_attributes.delete_if { |k, v| excluded_attributes.include? k }
 
         evaluation_attributes["evaluation_id"] = evaluation['id']
         evaluation_attributes["metadata_id"] = evaluation['metadata_id']
-        evaluation_attributes["url"] = evaluation['url']
-        evaluation_attributes["year"] = evaluation['year']
+        evaluation_attributes["url"] = evaluation['url'] || "N/A"
+        evaluation_attributes["year"] = evaluation['evaluation_year']
         evaluation_attributes["methodology"] = evaluation['methodology']
         evaluation_attributes["wdpa_id"] = evaluation['wdpa_id']
         evaluation_attributes["iso_3"] = evaluation['countries']
         evaluation_attributes["name"] = evaluation['site_name']
-        evaluation_attributes["designation"] = evaluation['designation']
+        evaluation_attributes["designation"] = evaluation['designation'] || "N/A"
         evaluation_attributes["source_data_title"] = evaluation['data_title']
         evaluation_attributes["source_resp_party"] = evaluation['resp_party']
         evaluation_attributes["source_year"] = evaluation['source_year']
@@ -250,18 +306,20 @@ class PameEvaluation < ApplicationRecord
     filter_params = json_params["_json"].nil? ? nil : json_params["_json"]
 
     where_statement = []
+    restricted_where_statement = []
     where_params = parse_filters(filter_params)
     where_params.map do |k, v|
       where_statement << where_fields(k,v) unless v.nil?
+      restricted_where_statement << where_fields(k,v) if !v.nil? && k != :sites
     end
 
     where_statement = where_statement.join(' AND ')
-    generate_csv(where_statement)
+    restricted_where_statement = restricted_where_statement.join(' AND ')
+    generate_csv(where_statement, restricted_where_statement)
   end
 
   def self.where_fields(k,v)
     return "e.#{v}" if [:year, :sites].include? (k)
     return v
-    generate_csv(where_params)
   end
 end